move all types to GenericButton
