+++ packages/components/src/Button/Button.tsx
@@ -1,30 +1,9 @@
 import React, { forwardRef, Ref } from "react"
-import GenericButton from "./components/GenericButton"
-import {
-  BadgeProps,
-  ButtonFormAttributes,
-  ButtonRef,
-  GenericProps,
-  WorkingProps,
-  WorkingUndefinedProps,
-} from "./types"
 
-export type ButtonProps = GenericProps &
-  ButtonFormAttributes &
-  (WorkingProps | WorkingUndefinedProps) & {
-    label: string
-    primary?: boolean
-    destructive?: boolean
-    secondary?: boolean
-    /** @default "regular" */
-    size?: "small" | "regular"
-    badge?: BadgeProps
-    type?: "submit" | "reset" | "button"
-    fullWidth?: boolean
-    iconPosition?: "start" | "end"
-    icon?: JSX.Element
-    disabled?: boolean
-  }
+import GenericButton, {
+  ButtonProps,
+  ButtonRef,
+} from "./components/GenericButton"
 
 /**
  * {@link https://cultureamp.design/components/button/ Guidance} |
