+++ packages/components/src/Button/index.ts
@@ -1,1 +1,2 @@
 export * from "./Button"
+export * from "./components/GenericButton"
